JSON serialize: support for DynamicObject + Improve JSON tests (#2902)
* JSON serialize: support for DynamicObject

* wipa fixup tests

* Fix logic

* refactor

* fix tests

* Split testcases for different interface impl
